chore: Port release infrastructure from ADBC (#99)
Closes #95.

I don't think this is *everything* that is needed for release but I
think it is close! There is still a few features that are needed in the
R package before a release candidate can go out, but I want to make sure
that enough of the infrastructure is in place that generating the
release candidate isn't the hold up.