fix(extensions/nanoarrow_ipc): Tweak draft interface (#111)
The initial draft interface had no way to retrieve the body length,
which made it more or less useless for iterating over streams without
decoding them (e.g., just verify a bunch of messages and get their
types). The initial interface also passed the input buffer by reference
and mutated it, but that wasn't really all that useful it turns out
(easier to provide the header + body lengths and let the caller advance
the input.