fix: make sure bundled nanoarrow is also valid C++ (#112)
Apparently it's difficult to pass different flags to C and C++ files in
a Python extension. The C library was almost valid C++ anyway and
because there are quite a few static inline functions, it's probably a
good idea to make sure that the project compiles under a C++ compiler as
well (since the nanoarrow.h header will be compiled with C++ if included
in a C++ project).