fix(extensions/nanoarrow_ipc): Don't release input stream automatically on end of input (#168)
Before this PR, the input was released from within `get_next()` in the
`ArrowArrayStream` implementation. This is unnecessary and prevented the
reader from being useful in the context of streams that have
non-standard layouts (e.g., if several streams are concatenated
together, one could continue calling `get_next()` to "try again"). Also,
there is a much better chance that `get_next()` is going to get called
on another thread and it's easier to reason about lifecycles if the
stream gets closed on `release()` instead.