feat!: Allow explicit validation level in `ArrowArrayFinishBuilding()` (#175)
This is a breaking change! The signature of `ArrowArrayFinishBuilding()`
gains a `validation_level` argument, and the original function was
renamed to `ArrowArrayFinishBuildingDefault()`. I could do this in a
non-breaking way but there are no non-awkward names for
`ArrowArrayFinishBuildingWithExplicitValidationLevel()` and it seems
more important to cultivate a quality API at this point rather than
maintain full backward compatibility (plus it's a relatively easy rename
fix).

This adds two important features. First, it lets you preform more
validation if you want to (via `NANOARROW_VALIDATION_LEVEL_FULL`,
implemented with the just-added `ArrowArrayViewValidateFull()`). Second,
it lets you perform less validation if you want to (e.g., if your
buffers are sitting on a GPU and dereferencing a buffer value will
crash).