chore(extensions/nanoarrow_ipc): Test files from the arrow-testing repo (#176)
This PR adds file tests for the integration files in
https://github.com/apache/arrow-testing/tree/master/data/arrow-ipc-stream/integration
(and fixed a few bugs identified as a result of reading them). The added
tests ensure that the data is read and that the data read is identical
to what Arrow C++ would read as well (or that files fail with a
particular error code and message).

I also tested the fuzzed files and didn't get any crashes; however I'd
like to spend more time on that and perhaps generate some more targeted
fuzzed files (for example, none of the fuzzed files tripped a failed
flatbuffer verification or a ValidateFull failure).

The Arrow IPC file tests run on CI, as a part of release verification,
and locally if you set the NANOARROW_ARROW_TESTING_DIR environment
variable to your favourite arrow-testing checkout.