chore(ci): Add documentation build tools to the Ubuntu docker image (#183)
The Ubuntu image is unofficially the "maximal" image (it also contains
the arrow R package). Adding this in a separate PR from the actual build
script so that I can manually trigger the image push once this is merged
(and then it will be available for pulling by CI in the relevant PR).