chore: Move bundling into its own workflow (#192)
Before, bundling was done as part of the test workflows and were
separate for the core and IPC extensions. This led to some problems when
a PR touched both since there was an attempt to commit a bundle to main
that was out of step. This PR moves that to a separate workflow that can
be dispatched manually in case this fails.