Add missing comma (#193)
Thanks for adding the nice example. While perusing it (and I need to
read it fully) I spotted a simple missing comma in the enumeration of
possible install commands. That's makes it possibly to most
insignificant PR ever but hey we believe in incremental fixes. Feel free
to close it if you'd rather address it elsewhere / as part of a larger
commit.