fix: Improve readability of `ArrowArrayAllocateChildren()` (#199)
In `ArrowArrayAllocateChildren()` two loops over all children follow
each other directly. The first assigns a NULL to each element, then
second overwrites that with an allocation. That makes the first loop
effectively deadweight. It may be there on purpose in which case you
just disregard and close this PR. But if it is an oversight this will
make it ever-so-marginally faster as we can skip one loop over
`n_children`.

(I noticed another microscopic style difference: in the corresponding
`ArrowSchemaAllocateChildren()` an `if (n_children != 0) {...}` wraps
around the core code of function; there it is a simpler `if (n_children
== 0) return ...`. Of course they are also the same "in practice" and I
don't have a style preference either -- just curious that they differ.
No need for change, but while at it you could....)

---------

Co-authored-by: Dewey Dunnington <dewey@fishandwhistle.net>