fix(r): Release streams when calling `as.vector()` or `as.data.frame()` (#202)
In https://github.com/apache/arrow-adbc/pull/706 , the added
`read_adbc()` releases a stream whose statement (and possibly
connection) will be released when the stream itself is released. When
the user is definitely consuming the entire stream, it is helpful to
release it at a guaranteed time so that subsequent operations (e.g.,
closing a connection, deleting a file on Windows) don't confusingly
fail.