refactor: Unify `ArrowArrayView` and `ArrowArray` validation (#201)
The initial motivation for this refactor was to make it possible to go
straight from Arrow IPC to ArrowArrayView with zero heap allocations and
improve IPC validation; however, the changes needed for that (basically,
give every ArrowArrayView its own offset, length, and null_count)
allowed unifying the validation that we do when building arrays and the
validation we do when wrapping a foregin ArrowArray in an
ArrowArrayView. This PR ensures that no internal ArrowArrayView
functions actually refer to the `array` member since there is no
guarantee that an ArrowArray backs the buffers. Along the way, error
messages and documentation were improved.