chore(ci): Add Python packaging tools to images + fix s390x build (#206)
s390x was failing because it was running `docker compose build` and
there are now some compose jobs that aren't intended to run on s390x
(e.g., documentation, coverage). Python build tools aren't essential yet
but will be at some point and updating the images is a bit of a pain.