feat: Add `enum ArrowType buffer_data_type` member to `struct ArrowLayout` (#207)
The existing `enum ArrowBufferType` is useful to switch on for some
things and the `element_size_bits` is useful for calculating byte
offsets; however, the combination of those is still not sufficient to do
endian-swapping as each Arrow type has its own rules. This concept has
also come up when printing buffers in the R package and when exporting
in Python via the buffer protocol, both of which had their own
workarounds. Attaching a reasonable `ArrowType` to each buffer should
provide a generic route to solving all of those.