refactor(extensions/nanoarrow_ipc): Reconfigure assembling arrays for better validation (#209)
Currently the decoder always assembles an `ArrowArray` and subsequently
validates that array. That caused a problem when validating a union
because the `ArrowArray` builder doesn't have a place to put custom type
IDs. Also, this is every so slightly wasteful: the `ArrowArray`
validation allocates an unnecessary `ArrowArrayView` and validates that.
We already have an `ArrowArrayView` to help with the buffer shuffling,
so this PR just swaps the direction: the first step is to create the
`ArrowArrayView`; the second step is to validate the `ArrowArrayView`,
then populate the `ArrowArray` (if requested).

Closes #178 by erroring if there is insufficient information to validate
a union array. I don't think this will cause problems with existing code
that builds a union array but if it does it could be relaxed to the
previous state (silently ignoring union arrays that can't be validated).