feat: Include dictionary member in `ArrowArrayView` struct (#221)
This doesn't add any features really, but ensures that one can walk
`ArrowSchema`, `ArrowArray`, and `ArrowArrayView` recursively using the
same pattern. I'd like to include this in the 0.2 release because it is
very difficult to work around this limitation: for example, if you have
a deeply nested dictionary field, you currently have to walk your whole
tree of arrays twice (once to validate the non-dictionary bits, once
looking for dictionary bits that may or may not exist to validate). The
R package worked around this in some creative but error-prone ways and
I'd like to avoid anybody else attempting creative workarounds when this
is a feature that will almost certainly get added soon.

Another application of this is device array support, since that PR
essentially constructs an `ArrowArrayView` and recursively copies the
buffer view from the array view to a standalone buffer using some
device-specific logic.