chore: Suppress more jemalloc thread local storage possible leaks and add memcheck for R to verification (#229)
As reported in #200, when loading the arrow R package and allocating at
least once there can be an additional thread local storage allocation by
jemalloc that leaks. I wasn't able to replicate this in Docker
(ubuntu:22.04 or debian:testing, which I believe is the platform CRAN
uses); however, their appearance shouldn't cause nanoarrow's memcheck to
fail (R or otherwise).

In trying to replicate this, I also realized that release verification
wasn't running the R tests with valgrind. On CI this isn't possible
because we install arrow from RSPM and that build crashes R under
valgrind for some reason. The verification job happens on Docker,
though, so we can add the check there.

Closes #200.