fix(extensions/nanoarrow_ipc): Check number of bytes read when reading buffer body (#295)
This PR adds a check in the `ArrowArrayStream` implementation when
reading the message body. The decoder checks for buffer overrun when
accessing specific buffers; however because message sizes are rounded up
to a multiple of 8 bytes, removing parts of a record batch message body
could lead to corrupted data being returned (even though no
out-of-bounds memory accesses occurred). This PR adds that check.