feat(r): Add support for `bit64::integer64()` conversions (#293)
This PR implements support for `bit64::integer64()`. This vctr type is a
mostly lossless representation of a 64-bit (signed) integer that is
helpful for representing very large numbers (or uses like `s2_cell()`
that just need a way to store 64 bits losslessly).

This type isn't part of the default conversions because it requires
`bit64`, which has a few dependencies. It is also not the default
because there are some limitations in R that cause that object type to
behave unintuitively in certain places (e.g., when indexing a vector).
Because `convert_vector()` always accepts a target vector type, users
can (now) override the default behaviour with a safer option if that
matters for their use case.

``` r
library(nanoarrow)
suppressPackageStartupMessages(library(bit64))

(vec <- as.integer64(c(-3:3, NA)))
#> integer64
#> [1] -3   -2   -1   0    1    2    3    <NA>
(array <- as_nanoarrow_array(vec))
#> <nanoarrow_array int64[8]>
#>  $ length    : int 8
#>  $ null_count: int 1
#>  $ offset    : int 0
#>  $ buffers   :List of 2
#>   ..$ :<nanoarrow_buffer validity<bool>[8][1 b]> `TRUE TRUE TRUE TRUE TRUE T...`
#>   ..$ :<nanoarrow_buffer data<int64>[8][64 b]> `-3.00e+00 -2.00e+00 -1.00e+0...`
#>  $ dictionary: NULL
#>  $ children  : list()
convert_array(array, integer64())
#> integer64
#> [1] -3   -2   -1   0    1    2    3    <NA>
```

<sup>Created on 2023-09-01 with [reprex
v2.0.2](https://reprex.tidyverse.org)</sup>