feat: Add integration testing JSON reader (#322)
This PR adds a `TestingJSONReader`. It uses
https://github.com/nlohmann/json to do the parsing. This doesn't
implement parsing a record batch or array yet, and there are a few
missing types (the same missing types as the writer). I'd like to do
those as a follow-up.

The testing for this is mostly roundtrip testing, which is maybe not
ideal but also maybe OK because the writer is an independent
implementation. When the ability to compare schemas for equality gets
added that can be tacked on to the roundtrip test utility function as
well.