chore(ci): Fix verification builds for Windows and centos7 (#341)
This PR contains a few updates such that the verification job passes on
Windows and centos7, and fedora.

On centos7/fedora, the builds were failing because the new test that was
added needs to be compiled with `-fPIC`. I updated the verification
script to always set `CMAKE_POSITION_INDEPENDENT_CODE=ON` from the
top-level `cmake` invocation. This was easier than trying to get
`target_compile_options()` working on the versions of cmake available by
default in all the runners.

On the GitHub Actions-based verifications, the builds were failing
because the cached Arrow C++ build was not built with the zlib/the gzip
codec. Simply adding `ARROW_WITH_ZLIB=ON` did not work on Windows
(something about a missing `.lib` file), so I just switched to using
zlib directly. When the IPC reader supports compression we will need it
anyway and it was a fairly minimal change.

There were also a few compiler warnings that I fixed while I had a real
live Windows/MSVC session going.