chore(ci): Update dockerfiles for improved R/Arrow C++ dependency management (#416)
Just a few quality-of-life improvements for the images to keep the R
dependencies automatically up-to-date.

I also checked that `xz` is not one of the affected versions...it had
been on the archlinux image at one point; however, rebuilding the image
several days ago puts the version at a version considered "fixed" by
archlinux ( https://security.archlinux.org/package/xz ). We don't use
this image for building any binaries that we distribute or otherwise
consume.