Increase CMake minimum version, move requirement (#422)
This is a very minor 'quality of life' improvement to `linesplitter` as
`cmake` nags about the minimum requirement being after (rather than
before) the main project declaration, and increasing the version (which
I understand you may not want to increase by as much) silences the
TIMESTAMP nag. Feel free to adjust or ignore, I just liked the quieter
build.