fix: Relax comparison strictness such that integration tests pass (#399)
These changes are the changes required such that
https://github.com/apache/arrow/pull/39302 results in passing
integration tests for nanoarrow. The changes are mostly related to
comparison:

- We needed an option to allow metadata to be compared on a key/value
basis without considering order (for Java, which seems to reorder
metadata on read)
- We needed the ability to treat NULL metadata and zero-size metadata as
equivalent (for Go, which always exports zero-length metadata)
- We needed an option to ignore flags for top-level batches (for C#,
which exports nullable structs)
- We needed to ensure that the last few bits of the validity buffer were
zeroed (for C#, although this is now fixed in C# on Arrow main)
- We needed to ensure that no buffers were NULL (For C#, which leaks the
top-level array if it encounters one, at least in the integration tests.
This should really be fixed in C#).