chore(python): Remove C sources from wheels (#447)
Even though the binaries were quite small, nanoarrow's wheels were about
1 MB and the installed size was about 5 MB. This is because the flatcc
headers and cython generated code was rather extensive and had ended up
in the built wheel. After this change the wheels are ~400kb and
installed size is 1.5 MB (for me).

The main change here is to move the directory where we vendor the C
files from elsewhere in the repo outside `src/` such they are not
automatically included in the wheel.