chore(python): Clean up top-level namespace (#449)
This PR cleans up the top-level namespace such that more advanced
concepts that might be confusing to new nanoarrow users are tucked away
in modules (i.e., they can be used and are documented, but better
alternatives are likely to be discovered first). The changes are:

- `allocate_c_array()`, `allocate_c_schema()`, and
`allocate_c_array_stream()` were removed. These are useful for testing
invalid input but pretty much nothing else.
- `c_array_view()` and `c_schema_view()` were removed. These are tools
used by the iterators, the `Array`, and the `Schema`, but most people
should use the higher level concepts for buffer access.
- A `schema()` alias for `Schema()` was added (because it is easy to
accidentally type `schema()` and get a wildly inappropriate error
message). This mirrors `array()` as an alias for `Array()`.