chore(dev/release): Do nanoarrow vendoring using R_BIN instead of R CMD INSTALL on Windows (#475)
It seems that running R CMD INSTALL is not sufficient on Windows to
vendor the proper files. It may have worked by accident in the past
because the vendored files already existed in my local checkout. I don't
think there are many people verifying on Windows; however I think
removing `configure.win` (since it seems not to run anyway) and
configure on build is a better option.