fix(python): Fix use of memoryview to write fill to the buffer builder (#477)
At least one test is failing on pypy because of an unreleased buffer.
This was due to the line:

```python
memoryview(builder)[out_start : out_start + length] = b"\x01" * length
```

...which relied on the deletion of `memoryview(builder)` to release the
`builder` buffer. One solution is:

```python
with memoryview(builder) as mv:
  mv[out_start : out_start + length] = b"\x01" * length
```

However, we also have `ArrowBufferAppendFill()` in the C library, so
instead, this PR wires that up and uses it (which should be faster,
anyway).