feat: Add support for appending values to half float `ArrowArray` (#499)
Hi, thanks for this great library! I found that `ArrowArrayAppendXX`
functions won't append a value to a half float `ArrowArray` while I was
playing with this library. So I decided to add this support, which I
believe it should make this library easier to use.

The conversion function is adapted from Arrow Go,
https://github.com/apache/arrow/blob/main/go/arrow/float16/float16.go,
and the tests cases are taken from
https://github.com/apache/arrow/blob/main/go/arrow/float16/float16_test.go.

Although I'm not sure if I should also update the files in `dist` though
(I guess it might be automated when doing a release but I didn't check).
I'll be happy to update them if these files are not updated
automatically, and I'm happy to do any other changes. :)