refactor(r): Use JSON in experimental R vctrs extension type (#533)
This PR updates the "vctrs extension type" to use JSON as its metadata
serialization. JSON is use for most other extension types and using it
here provides some workaround for non-R consumers that encounter this
type elsewhere.

The serializer and deserializer use the same format as
`jsonlite::serializeJSON()`, but restrict the types of objects that it
is willing to serialize or deserialize.