fix(ci): Pin r-lib actions as a workaround for latest action updates (#572)
First noticed at
https://github.com/apache/arrow-nanoarrow/pull/555#discussion_r1702201173
, the R check action is failing because an update to r-lib actions
resulted in some quarto actions being invoked, and these have not yet
been whitelisted for use in Apache repositories. It also may be that we
don't need the quarto actions (we probably don't) but some brief
experimentation to attempt circumventing the use of the quarto action
did not result in a successful workflow. Hence, a pin to unblock PR
checks until either the v2 branch is updated or it is clear how to avoid
the failure.