feat: Add `ArrowArrayViewCompare()` to check for array equality (#578)
This PR is one possible component to address #577. While in some cases
we want a more relaxed comparison that allows (for example) arrays with
the same content to be considered equal even if they have different
content in null slots, in some cases we really do want an exact match.
This PR adds `ArrowArrayViewCompare()` in such a way that the same
signature could be used to apply the equality check at a more relaxed
validation level when this is implemented in a future PR, but only
implements the "identical" level since this is the easiest/most pressing
(applies to IPC validation).

The messages given by the implementation give the location of the
difference but not what the difference actually was. Knowing where the
error was is usually sufficient for a higher level runtime (e.g., R,
Python, C++) to give a fancier message if they want or need to.